api/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
PyJWT>=2.8,<3
pytest>=8,<9
httpx>=0.27,<1
